// apps/navigator/src/models/automations/automations.ts
export interface AutomationResource {
  id: string;
  name: string;
  datamart_id: string;
  organisation_id: string;
  status: AutomationStatus;
}

export function isAutomationResource(
  automation: AutomationSimpleFormData,
): automation is AutomationResource {
  return (automation as AutomationResource).id !== undefined;
}

export type AutomationStatus = 'NEW' | 'ACTIVE' | 'PAUSED';